Figure: HTML
Reader: `<figure>` elements are parsed as figures, with the caption
taken from the respective `<figcaption>` elements.

Writer: the alt text is no longer constructed from the caption, as was
the case with implicit figures. This reduces duplication, but comes at
the risk of missing alt texts. Authors should take care to provide alt
texts for all images.