Figure: LaTeX
Reader: support for figures with non-image contents and for subfigures.

Writer: complex figures, e.g. with non-image contents and subfigures,
are supported. The `subfigure` template variable is set if the document
contains subfigures, triggering the conditional loading of
the *subcaption* package. Contants of figures that contain tables are
become unwrapped, as longtable environments are not allowed within
figures.